// src/app/interfaces/enum.ts
export enum teachers {
  ivanov = 'Олег Иванов',
  lavrov = 'Михаил Лавров',
  malakhova = 'Олеся Малахова',
  kapura = 'Елена Капура',
  lyapcev = 'Евгений Ляпцев',
  lyapceva = 'Екатерина Ляпцева',
  lavrenova = 'Александра Лавренова',
  stepanova = 'Анастасия Степанова',
  galaeva = 'Лейла Галаева',
  vetkova = 'Веткова Анастасия',
  vasilenko = 'Анна Василенко',
  naimushin = 'Алексей Наймушин',
  boiko = 'Ольга Бойко',
  philatova = 'Ольга Филатова',
  antonov = 'Сергей Антонов',
  nevolin = 'Борис Неволин',
  glazkova = 'Юлия Глазкова',
}

export enum PageUrlEnum {
  products = 'products',
  myProducts = 'my-products',
  game = 'game',
}